==> Performance_benchmarks/SingleCell/Tasic_dataset/2_Tasic_DTU.Rmd <==
---
title: "2_Tasic_DTU"
author: "Jeroen Gilis"
date: "05/11/2020"
output: html_document
---

**In order to run this script (2_Tasic_DTU.Rmd), the dataset Tasic_benchmark_datasets_count.Rdata (or, alternatively, Tasic_benchmark_datasets_scaledTPM.Rdata) is required.** This file can either be generated with the 1_Tasic_prepare.Rmd script or downloaded from Zenodo.

Here we run the DTU analyses for all 6 methods on all (18) Tasic benchmark datasets. Note that for DEXSeq and DRIMSeq we only run the datasets with 20 cells in each group, as these methods do not scale to large datasets. NBSplice was omitted as it does not converge on datasets with many zeroes. This code runs approximately 2.5 hours on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of this runtime was attributed to the DEXSeq and DRIMSeq analyses.

**If you do not want to run this script, its output can also be downloaded from Zenodo: Tasic_DTU_results_count.Rdata (or, alternatively, Tasic_DTU_results_scaledTPM.Rdata) **

```{r setup, include=FALSE, echo=FALSE}
require("knitr")
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Load DTU methods

Source file that allows running all seven DTU methods that were assessed in this paper: satuRn, DoubleExpSeq, limma diffsplice, edgeRdiffsplice, DEXSeq, DRIMseq and NBSplice.

```{r, message=FALSE, warning=FALSE}
### load libraries
library(edgeR)
library(limma)
library(DEXSeq)
library(DRIMSeq)
library(DoubleExpSeq)
library(NBSplice)
library(satuRn)
library(doMC)

source(file="./Performance_benchmarks/DTU_methods.R")
```

# Set up parallel execution

Note that the optimal parameter setting for parallelizing the DTU analyses depends on your specific computing system.

```{r}
if(TRUE) {
    nrCores <- 2

    if(nrCores != 1) {
        doParallel <- TRUE
        doProgress <- 'none'

        registerDoMC(cores = nrCores)
    } else {
        doParallel <- FALSE
        doProgress <- 'text'
    }
}
```

# Run all analyses 

Run the DTU analyses for all 6 methods on all (18) Chen benchmark datasets. Note that for DEXSeq and DRIMSeq we only run the datasets with 20 cells in each group, as these methods do not scale to large datasets. NBSplice was omitted as it does not converge on datasets with many zeroes.
This code runs approximately 3 hours on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of this runtimee was attributed to the DEXSeq and DRIMSeq analyses.

```{r}
### Load benchmark data
load(file="./Data/Tasic_benchmark_datasets_count.Rdata")
#load(file="./Data/Tasic_benchmark_datasets_scaledTPM.Rdata")

TasicBenchmarkData <- c(TasicBenchmarkLenient,TasicBenchmarkStringent)

### Run DTU analysis on benchmark data

print("start satuRn")

tStart <- Sys.time()
suppressWarnings(TasicDtuBenchmark_satuRn <- plyr::llply(
    .data = TasicBenchmarkData,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- satuRn_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
))
difftime(Sys.time(), tStart)

print("start limma diffsplice")

tStart <- Sys.time()
TasicDtuBenchmark_limmaDiffsplice <- plyr::llply(
    .data = TasicBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- limma_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes) )
    }
)
difftime(Sys.time(), tStart)

print("start edgeR_diffsplice")

tStart <- Sys.time()
TasicDtuBenchmark_edgeRdiffsplice <- plyr::llply(
    .data = TasicBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- edgeR_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]


        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start DoubleExpSeq")

tStart <- Sys.time()
TasicDtuBenchmark_DoubleExpSeq <- plyr::llply(
    .data = TasicBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- DoubleExpSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)


print("start DEXSeq")

# DEXSeq
tStart <- Sys.time()
TasicDtuBenchmark_DEXSeq <- plyr::llply(
    .data = TasicBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        
        if(ncol(localData$data) > 40) {
            return(NULL)
        }
      
        ### Perform DTU analysis
        localRes <- DEXSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start DRIMSeq")

tStart <- Sys.time()
TasicDtuBenchmark_DRIMSeq <- plyr::llply(
    .data = TasicBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
      
        if(ncol(localData$data) > 40) {
            return(NULL)
        }
      
        ### Perform DTU analysis
        localRes <- DRIMSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

### add method name to list names for easy post-analysis
names(TasicDtuBenchmark_satuRn) <- paste0('satuRn_', names(TasicDtuBenchmark_satuRn))
names(TasicDtuBenchmark_limmaDiffsplice) <- paste0('limma_diffsplice_', names(TasicDtuBenchmark_limmaDiffsplice))
names(TasicDtuBenchmark_DEXSeq) <- paste0('DEXSeq_', names(TasicDtuBenchmark_DEXSeq))
names(TasicDtuBenchmark_DRIMSeq) <- paste0('DRIMSeq_', names(TasicDtuBenchmark_DRIMSeq))
names(TasicDtuBenchmark_edgeRdiffsplice) <- paste0('edgeR_diffsplice_'        , names(TasicDtuBenchmark_edgeRdiffsplice))
names(TasicDtuBenchmark_DoubleExpSeq) <- paste0('DoubleExpSeq_', names(TasicDtuBenchmark_DoubleExpSeq))

### Save result
save(TasicDtuBenchmark_satuRn,
    TasicDtuBenchmark_limmaDiffsplice,
    TasicDtuBenchmark_DEXSeq,
    TasicDtuBenchmark_DRIMSeq,
    TasicDtuBenchmark_edgeRdiffsplice,
    TasicDtuBenchmark_DoubleExpSeq,
    file="./Data/Tasic_DTU_results_count.Rdata")

# save(TasicDtuBenchmark_satuRn,
#     TasicDtuBenchmark_limmaDiffsplice,
#     TasicDtuBenchmark_DEXSeq,
#     TasicDtuBenchmark_DRIMSeq,
#     TasicDtuBenchmark_edgeRdiffsplice,
#     TasicDtuBenchmark_DoubleExpSeq,
#     file="./Data/Tasic_DTU_results_scaledTPM.Rdata")
```
