==> Performance_benchmarks/SingleCell/Tasic_dataset/3_Tasic_visualize.Rmd <==
---
title: "3_Tasic_visualize"
author: "Jeroen Gilis"
date: "06/11/2020"
output: html_document
---

**In order to run this script (3_Tasic_visualize.Rmd), the datasets Tasic_benchmark_datasets_count.Rdata and Tasic_DTU_results_count.Rdata (or, alternatively, Tasic_benchmark_datasets_scaledTPM.Rdata and Tasic_DTU_results_scaledTPM.Rdata) are required.** These files can either be generated with the 1_Tasic_prepare.Rmd and 2_Tasic_DTU.Rmd scripts, respectively, or downloaded from Zenodo  and put in the data folder of this GitHub repository.

The results of this script (FDR-TPR curves) are already available from our GitHub, under the directory ./Results/Tasic_benchmark/.

```{r setup, include=FALSE, echo=FALSE}
require("knitr")
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Load libraries

```{r,message=FALSE,warning=FALSE}
library(ggplot2)
library(iCOBRA)
```

# Load truth file

```{r}
load(file="./Data/Tasic_benchmark_datasets_count.Rdata")
#load(file="./Data/Tasic_benchmark_datasets_scaledTPM.Rdata")

metaInfo <- c(TasicBenchmarkLenient,TasicBenchmarkStringent)

rm(TasicBenchmarkLenient,TasicBenchmarkStringent)
```

# Load DTU performance data

```{r}
load(file="./Data/Tasic_DTU_results_count.Rdata")
#load(file="./Data/Tasic_DTU_results_scaledTPM.Rdata") # to run the visualization on the scaledTPM data

TasicBenchmark <- c(
    TasicDtuBenchmark_satuRn,
    TasicDtuBenchmark_limmaDiffsplice,
    TasicDtuBenchmark_DEXSeq,
    TasicDtuBenchmark_DRIMSeq,
    TasicDtuBenchmark_edgeRdiffsplice,
    TasicDtuBenchmark_DoubleExpSeq)

### Remove empty enteries (due to not tested - aka to many samples for reasonable runtime)
TasicBenchmark  <- TasicBenchmark[which(
    sapply(TasicBenchmark, function(x) ! is.null(x$dtuAnalysis)))]

rm(TasicDtuBenchmark_satuRn,
    TasicDtuBenchmark_limmaDiffsplice,
    TasicDtuBenchmark_DEXSeq,
    TasicDtuBenchmark_DRIMSeq,
    TasicDtuBenchmark_edgeRdiffsplice,
    TasicDtuBenchmark_DoubleExpSeq)
invisible(gc())
```

# Helper function 
To get pvalues for the different methods on the different datasets

```{r}
get_pvalues <- function(dataset,nrep,nmethods) {
  
  get_all_txs <- function(dataset,nrep){
  
    all_txs <- c()
    for (i in 1:nrep) {
      current_txs <- as.character(dataset[[i]]$dtuAnalysis$TXNAME)
      all_txs <- c(all_txs, current_txs)
      all_txs <- unique(all_txs)
    }
    return(all_txs)
  }

  all_txs <- get_all_txs(dataset,nrep)
  pvalues <- matrix(data=NA, nrow=length(all_txs), ncol=nmethods*nrep)
  rownames(pvalues) <- all_txs
  
  for (i in 1:(nmethods*nrep)) {
    pvalues_new <- data.frame(dataset[[i]]$dtuAnalysis$p_value)
    rownames(pvalues_new) <- as.character(dataset[[i]]$dtuAnalysis$TXNAME)
    colnames(pvalues_new) <- names(dataset)[i]

    pvalues[,i] <- pvalues_new[match(rownames(pvalues),rownames(pvalues_new)),1]
  }
  
  pvalues <- apply(pvalues, 2, as.numeric)
  rownames(pvalues) <- all_txs
  pvalues <- as.data.frame(pvalues)
  pvalues <- pvalues[order(rownames(pvalues)),]
  
  if(nmethods == 6){
    colnames(pvalues) <- c("satuRn", "limma_diffsplice", "DEXSeq", "DRIMSeq","edgeR_diffsplice", "DoubleExpSeq")
  } else {
    colnames(pvalues) <- c("satuRn", "limma_diffsplice","edgeR_diffsplice", "DoubleExpSeq")
  }
  return(pvalues)
}
```

# Visualization function

```{r}
visualize_datasets <- function(dataset,metaInfo,nFilters,nSampleSizes,nRep,nmethods,selection,selection2) {

    count <- 0
    nPlots <- nFilters * nSampleSizes
    print(paste("This function will generate",  nPlots, "plots"))
  
    for (j in 1:nPlots) {
        print(j)
        Tasic_current_outer <- dataset[grepl(selection[j], names(dataset))]
        metaInfo_current_outer <- metaInfo[grepl(selection2[j], names(metaInfo))]

        resList <- c()
        count <- count+1

        for (k in 1:length(nRep)) {
            Tasic_current_inner <- Tasic_current_outer[grepl(nRep[k], names(Tasic_current_outer))]

            metaInfo_current_inner <- metaInfo_current_outer[grepl(nRep[k], names(metaInfo_current_outer))]
    
            pvalues <- get_pvalues(Tasic_current_inner,nrep=1,nmethods)
            #pvalues <- pvalues[,c(2,3)]

            truth_file <- Tasic_current_inner[[1]]$dtuAnalysis
    
            if(count <= nSampleSizes){
                txSwapped <- metaInfo_current_inner[1]          
                } else {
                txSwapped <- metaInfo_current_inner[2]
            }
    
            truth_file <- truth_file[,c("TXNAME","p_value")]
    
            txSwapped[[1]]$metaInfo <- txSwapped[[1]]$metaInfo[match(truth_file$TXNAME,txSwapped[[1]]$metaInfo$TXNAME),]
    
            truth_file$gene_modified <- as.numeric(txSwapped[[1]]$metaInfo$txSwapped)
            truth_file <- truth_file[order(truth_file$TXNAME),]
            rownames(truth_file) <- truth_file$TXNAME

            resList[[k]] <- list(pvalues,truth_file)
        }
        
        rownames(resList[[1]][[1]]) <- paste0(rownames(resList[[1]][[1]]), "_1")
        rownames(resList[[2]][[1]]) <- paste0(rownames(resList[[2]][[1]]), "_2")
        rownames(resList[[3]][[1]]) <- paste0(rownames(resList[[3]][[1]]), "_3")
        
        pvalues_full <- rbind(resList[[1]][[1]],resList[[2]][[1]],resList[[3]][[1]])
        
        rownames(resList[[1]][[2]]) <- paste0(rownames(resList[[1]][[2]]), "_1")
        rownames(resList[[2]][[2]]) <- paste0(rownames(resList[[2]][[2]]), "_2")
        rownames(resList[[3]][[2]]) <- paste0(rownames(resList[[3]][[2]]), "_3")
        
        truth_file_full <- rbind(resList[[1]][[2]],resList[[2]][[2]],resList[[3]][[2]])
        
        cobra <- COBRAData(pval = pvalues_full, truth = truth_file_full)
        cobra <- calculate_adjp(cobra)
        cobra1perf <- calculate_performance(cobra, binary_truth = "gene_modified", cont_truth = "none", splv = "none", aspects = c("fdrtpr", "fdrtprcurve", "overlap"))

        cobraplot <- prepare_data_for_plot(cobra1perf, colorscheme = "Dark2", facetted = TRUE)
        
        if (nmethods == 6){
            new_col <- c(rep(c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "black", "#999999"),4),rep("white",14))
            names(new_col) <- names(cobraplot@plotcolors)
            cobraplot@plotcolors <- new_col
    
            plot <- plot_fdrtprcurve(cobraplot, xaxisrange = c(0, 0.4), yaxisrange = c(0,0.75))
    
            titles <- c("20 versus 20 - edgeR filter - count ","20 versus 20 - DRIMSeq filter - count")
            #titles <- c("20 versus 20 - edgeR filter - scaledTPM ","20 versus 20 - DRIMSeq filter - scaledTPM")
            current_title <- titles[j]
        } else {
            new_col <- c(rep(c("#56B4E9", "#F0E442", "#0072B2", "black", "#999999"),4),rep("white",10))
            names(new_col) <- names(cobraplot@plotcolors)
            cobraplot@plotcolors <- new_col
    
            plot <- plot_fdrtprcurve(cobraplot, xaxisrange = c(0, 0.3), yaxisrange = c(0,1))
    
            titles <- c("75 versus 75 - edgeR filter - count","200 versus 200 - edgeR filter - count","75 versus 75 - DRIMSeq filter - count","200 versus 200 - DRIMSeq filter - count")
            #titles <- c("75 versus 75 - edgeR filter - scaledTPM","200 versus 200 - edgeR filter - scaledTPM","75 versus 75 - DRIMSeq filter - scaledTPM","200 versus 200 - DRIMSeq filter - scaledTPM")
            current_title <- titles[j]
        }

        plot <- plot + 
            ggtitle(current_title) +
            theme(strip.background = element_blank(),
            strip.text.x = element_blank(),
            plot.title = element_text(size=10),
            axis.text.x = element_text(size = 10),
            axis.text.y = element_text(size = 10),
            axis.title.x = element_text(size = 10),
            axis.title.y = element_text(size = 10))

      png(paste("./Results/Tasic_benchmark/FDRTPR_", selection2[j], "_", sub("^[^f]*", "", selection[j]), "_count.png", sep = ""),
            width     = 5.5,
            height    = 4.5,
            units     = "in",
            res       = 200,
            pointsize = 4) # start export
      print(plot)
      dev.off()
      
      # png(paste("./Results/Tasic_benchmark/FDRTPR_", selection2[j], "_", sub("^[^f]*", "", selection[j]), "_scaledTPM.png", sep = ""),
      #       width     = 5.5,
      #       height    = 4.5,
      #       units     = "in",
      #       res       = 200,
      #       pointsize = 4) # start export
      # print(plot)
      # dev.off()
  }
}
```

# Generate FDR-TPR curves

```{r}
selection <- c("20_rep_._filterLenient","20_rep_._filterStringent")
selection2 <- c("20_rep","20_rep")

visualize_datasets(dataset=TasicBenchmark,metaInfo=metaInfo,nFilters=2,nSampleSizes=1,nRep=c("rep_1","rep_2","rep_3"),nmethods=6,selection = selection,selection2=selection2)

selection <- c("_75_rep_._filterLenient","200_rep_._filterLenient","_75_rep_._filterStringent","200_rep_._filterStringent")
selection2 <- c("75_rep","200_rep","75_rep","200_rep")

visualize_datasets(TasicBenchmark,metaInfo,nFilters=2,nSampleSizes=2,nRep=c("rep_1","rep_2","rep_3"),nmethods=4,selection = selection,selection2=selection2)
```
